+++ src/units/udisplay.cpp
@@ -657,7 +657,9 @@ void unit_attack(display * disp, game_board & board,
 
 	animator.add_animation(defender.shared_from_this(), defender_anim, def->get_location(), true, text, {255, 0, 0});
 
-	for(const unit_ability& ability : attacker.get_abilities_weapons("leadership", weapon, secondary_attack)) {
+	unit_ability_list leadership_list = attacker.get_abilities_weapons("leadership", weapon, secondary_attack);
+	unit_ability_list resistance_list = defender.get_abilities_weapons("resistance", secondary_attack, weapon);
+	for(const unit_ability& ability : leadership_list) {
 		if(ability.teacher_loc == a) {
 			continue;
 		}
@@ -674,7 +676,7 @@ void unit_attack(display * disp, game_board & board,
 			hit_type, weapon, secondary_attack, swing);
 	}
 
-	for(const unit_ability& ability : defender.get_abilities_weapons("resistance", secondary_attack, weapon)) {
+	for(const unit_ability& ability : resistance_list) {
 		if(ability.teacher_loc == a) {
 			continue;
 		}
@@ -688,6 +690,49 @@ void unit_attack(display * disp, game_board & board,
 		helper->set_facing(ability.teacher_loc.get_relative_dir(b));
 		animator.add_animation(helper.get_shared_ptr(), "resistance", ability.teacher_loc,
 			def->get_location(), damage, true,  "", {0,0,0},
+			hit_type, weapon, secondary_attack, swing);
+	}
+
+	unit_ability_list abilities = att->get_location();
+	for(auto& special : attacker.checking_tags()) {
+		abilities.append(weapon->list_ability(special));
+	}
+
+	for(const unit_ability& ability : abilities) {
+		if(ability.teacher_loc == a) {
+			continue;
+		}
+
+		if(ability.teacher_loc == b) {
+			continue;
+		}
+
+		bool anim_playable = true;
+		for(const unit_ability& leader_list : leadership_list) {
+			if(ability.teacher_loc == leader_list.teacher_loc) {
+				anim_playable = false;
+				break;
+			}
+		}
+		if(!anim_playable) {
+			continue;
+		}
+
+		for(const unit_ability& helper_list : resistance_list) {
+			if(ability.teacher_loc == helper_list.teacher_loc) {
+				anim_playable = false;
+				break;
+			}
+		}
+		if(!anim_playable) {
+			continue;
+		}
+
+		unit_map::const_iterator leader = board.units().find(ability.teacher_loc);
+		assert(leader.valid());
+		leader->set_facing(ability.teacher_loc.get_relative_dir(a));
+		animator.add_animation(leader.get_shared_ptr(), "special", ability.teacher_loc,
+			att->get_location(), damage, true,  "", {0,0,0},
 			hit_type, weapon, secondary_attack, swing);
 	}
 
@@ -725,7 +770,11 @@ void reset_helpers(const unit *attacker,const unit *defender)
 	display* disp = display::get_singleton();
 	const unit_map& units = disp->get_units();
 	if(attacker) {
-		for(const unit_ability& ability : attacker->get_abilities("leadership")) {
+		unit_ability_list attacker_abilities = attacker->get_abilities("leadership");
+		for(auto& special : attacker->checking_tags()) {
+			attacker_abilities.append(attacker->get_abilities(special));
+		}
+		for(const unit_ability& ability : attacker_abilities) {
 			unit_map::const_iterator leader = units.find(ability.teacher_loc);
 			assert(leader != units.end());
 			leader->anim_comp().set_standing();
@@ -733,7 +782,11 @@ void reset_helpers(const unit *attacker,const unit *defender)
 	}
 
 	if(defender) {
-		for(const unit_ability& ability : defender->get_abilities("resistance")) {
+		unit_ability_list defender_abilities = defender->get_abilities("resistance");
+		for(auto& special : defender->checking_tags()) {
+			defender_abilities.append(defender->get_abilities(special));
+		}
+		for(const unit_ability& ability : defender_abilities) {
 			unit_map::const_iterator helper = units.find(ability.teacher_loc);
 			assert(helper != units.end());
 			helper->anim_comp().set_standing();